requirements.txt
diffusers
invisible_watermark
transformers
accelerate
safetensors
pika
imagekitio
python-dotenv
compel